Update CRUD Products

+++ backend/BackEnd/src/main/java/com/example/backend/Repository/ProductsRepository.java
@@ -3,5 +3,5 @@ package com.example.backend.Repository;
 import com.example.backend.Entity.Products;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductsRepository extends JpaRepository<Products,Integer> {
+public interface ProductsRepository extends JpaRepository<Products,Long> {
 }
